+++ src/admin/auction-item.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace gardenClubOfMpls\CustomFunctionalityPlugin\Source\Admin;
+
+// Add code here to render CPT post metadata in admin columns and make columns sortable